requirements.txt
openai>=1.6.0
python-dotenv==1.0.0
splitwise==2.4.0
requests==2.31.0
pytest
